opc.js; working remote setup
Lots of stuff! Introduce opc.js, a p5js library. Hoping it'll be pretty
reusable. Currently pulls off the first 400 pixels from the pixels array
and sends them along.
And running SimpleHTTPServer along with everything else, this works!!
I can load it on my phone and control the gl_server's pixels, which is
basically an end-to-end test. Add some stuff to try to get it app-like,
without scrollbars and such. Tries to run fullscreen. Etc.
A pretty exciting proof of concept!