Replay!
This makes this thing actually usable for a party: a replay feature that
will replay all old touches so there's something to look at when no
one's touching the thing. It's all rather convoluted but not a whole lot
of code in practice, and it's well-commented. Picks a random decay value
when replay mode starts; picks a random diameter/color for each touch.
Next: better color options!