Stop redrawing whole canvas on each draw()
It doesn't seem to be faster on my iPad, but does seem to dramatically
increase framerate on my OSX Chrome. This insight came from trying to
debug this:
https://github.com/processing/p5.js/issues/1485
It seems better to just invert / un-invert the pixels we need to, rather
than save/restore the whole canvas. Could that be contributing to the
memory usage?