Use overlay canvas to show pixel indicators
This is much faster and cleaner than doing weird stuff to draw/erase
pixel indicators on the main canvas. (I was actually doing it wrong too,
since the very first preDraw() was inverting black pixels!)
This should make everything faster and more reliable, plus the pixel
indicators are bigger and have the added benefit of not obscuring the
actual pixels being rendered to the display. Wins all around.