Basic working wifi serving!
This worked at a Brickolage deployment; wow!!
I really need to break things out into separate files, and I'm not
totally sure all this is necessary, and I haven't tried this on a fresh
system, but it works!! creates a "Brickolage" wifi network; a tablet can
connect and visit 192.168.1.1 in the browser and they'll see the
Brickolage interface. fcserver's running, detects plugged-in usb
devices, and works too!

Still TODO: make the network open; make random browser visits redirect
to http://brickolage/ which has the main UI. There should be a way to
visit :7890 for the fcserver web UI too.

Most of the instructions for this stuff came from this post:
https://frillip.com/using-your-raspberry-pi-3-as-a-wifi-access-point-with-hostapd/